requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
loguru>=0.6.0
tensorboard>=2.10.0
argparse
json5
typing-extensions
tqdm>=4.64.0 